flask/requirements.txt
ultralytics==8.2.0
opencv-python-headless==4.9.0.80
easyocr==1.7.1
numpy==1.26.4
pillow==10.2.0